Changed script to create symlinks based on files in directory

+++ makesymlinks.sh
@@ -8,26 +8,26 @@
 
 dir=~/dotfiles                    # dotfiles directory
 olddir=~/dotfiles.bak             # old dotfiles backup directory
-# list of files/folders to symlink in homedir
-files="bash_profile gitconfig gitignore_global hgignore_global mrconfig vimrc zshrc"
-
 ##########
 
 # create dotfiles_old in homedir
-echo "Creating $olddir for backup of any existing dotfiles in ~"
-mkdir -p $olddir
-echo "...done"
-
-# change to the dotfiles directory
-echo "Changing to the $dir directory"
+if [[ ! -d $olddir ]]; then
+    echo "Creating $olddir for backup of any existing dotfiles in ~"
+    mkdir -p $olddir
+fi
 cd $dir
-echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file $olddir/
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    echo .$file
+    if [[ -f ~/.$file && ! -L ~/.$file ]]; then
+        echo "Moving ~/.$file to $olddir..."
+        mv ~/.$file $olddir/
+        echo "Creating symlink to $file..."
+        ln -s $dir/$file ~/.$file
+    else
+        echo "File already exists as symlink."
+    fi
+    echo ''
 done
